+++ build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 }
 
 application {
-    mainClassName = "com.github.vacxe.googleplaycli.ApkPublisherKt"
+    mainClassName = "com.github.vacxe.googleplaycli.PlayStoreCliKt"
     applicationName = "google-play-cli"
 }
