--- src/main/kotlin/com/github/vacxe/googleplaycli/core/Extention.kt
@@ -1,2 +1,0 @@
-package com.github.vacxe.googleplaycli.core
-
