+++ src/main/kotlin/com/github/vacxe/googleplaycli/dsl/dsl.kt
@@ -2,7 +2,7 @@ package com.github.vacxe.googleplaycli.dsl
 
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.subcommands
-import com.github.vacxe.googleplaycli.PlayStoreCli
+import com.github.vacxe.googleplaycli.PlayStoreApi
 import com.github.vacxe.googleplaycli.core.BaseCommand
 import com.github.vacxe.googleplaycli.core.MetaCommand
 
@@ -12,10 +12,10 @@ fun MetaCommand.subcmd(name: String, configurationBlock: MetaCommand.() -> Unit)
     subcommands(cmd(name, configurationBlock))
 }
 
-fun MetaCommand.baseCmd(name: String, description: String = "", block: BaseCommand.(PlayStoreCli) -> Any) {
+fun MetaCommand.baseCmd(name: String, description: String = "", block: BaseCommand.(PlayStoreApi) -> Any) {
     subcommands(
             object : BaseCommand(name = name, actionDescription = description) {
-                override fun run(cli: PlayStoreCli) = block(cli)
+                override fun run(api: PlayStoreApi) = block(api)
             }
     )
 }
